Set cbreak when starting
From nodelay(3)

Initially the terminal may or may not be in cbreak mode, as the mode is
inherited; therefore, a program should call cbreak or nocbreak
explicitly. Most interactive programs using curses set the cbreak mode.
Note that cbreak overrides raw. [See curs_getch(3X) for a discussion of
how these routines interact with echo and noecho.]

This allows sl to function perfectly on NetBSD Curses.